Fix package naming
