+++ Examples/HelloTabWidget/src/course/examples/helloworldtabwidget/ArtistsActivity.java
@@ -1,15 +1,15 @@
-package course.examples.HelloWorldTabWidget;
+package course.examples.helloworldtabwidget;
 
 import android.app.Activity;
 import android.os.Bundle;
 import android.widget.TextView;
 
-public class AlbumsActivity extends Activity {
+public class ArtistsActivity extends Activity {
 	 public void onCreate(Bundle savedInstanceState) {
 	        super.onCreate(savedInstanceState);
 
 	        TextView textview = new TextView(this);
-	        textview.setText("This is the Albums tab");
+	        textview.setText("This is the Artists tab");
 	        setContentView(textview);
 	    }
 }
